Move tooltip text to constants, refactor HeaderList
