--- src/pages/nftLending/LendPage/PlaceOffersContent/components/HeaderList/index.ts
@@ -1,1 +1,0 @@
-export { default } from './HeaderList'
