--- src/pages/nftLending/LendPage/PlaceOffersContent/components/MarketOverviewInfo/MarketOverviewInfo.module.less
@@ -1,106 +1,0 @@
-// * MarketMainInfo * //
-.mainInfoContainer {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-
-  max-width: 290px;
-  width: 100%;
-
-  @media (min-width: 961px) and (max-width: 1240px), (max-width: 640px) {
-    max-width: 176px;
-  }
-
-  @media (max-width: 960px) {
-    gap: 8px;
-  }
-
-  svg {
-    width: 16px;
-    height: 16px;
-  }
-}
-
-.collectionImage {
-  border-radius: 100%;
-  height: 22px;
-  width: 22px;
-}
-
-.collectionName {
-  font: var(--title-md);
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-}
-
-.tensorLink {
-  svg path {
-    fill: var(--content-secondary);
-  }
-
-  &:hover {
-    svg path {
-      fill: var(--content-primary);
-    }
-  }
-}
-// * MarketMainInfo * //
-
-// * MarketAdditionalInfo * //
-.additionalInfoStats {
-  display: flex;
-  justify-content: flex-end;
-  gap: 16px;
-  white-space: nowrap;
-  width: 100%;
-
-  @media (max-width: 960px) {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    grid-template-rows: repeat(2, auto);
-    gap: 8px;
-  }
-}
-
-.additionalInfoStat {
-  align-items: flex-end;
-  justify-content: center;
-  gap: 2px;
-  max-width: 180px;
-  width: 100%;
-
-  @media (max-width: 1240px) {
-    max-width: 120px;
-  }
-
-  @media (max-width: 960px) {
-    .opened &:nth-child(n + 3) {
-      visibility: hidden;
-      height: 0;
-    }
-
-    &:nth-child(-n + 2) {
-      grid-row: 1;
-    }
-
-    &:nth-child(n + 3) {
-      grid-row: 2;
-    }
-
-    &:not(:last-child) {
-      align-items: flex-start;
-    }
-  }
-}
-
-.additionalInfoStatLabelWrapper {
-  @media (min-width: 961px) {
-    display: none;
-  }
-}
-
-.additionalAprStat {
-  color: var(--additional-green-primary-deep);
-}
-// * MarketAdditionalInfo * //
