--- src/pages/nftLending/LendPage/PlaceOffersContent/components/MarketOverviewInfo/index.ts
@@ -1,1 +1,0 @@
-export * from './MarketOverviewInfo'
