+++ src/pages/nftLending/LendPage/PlaceOffersContent/constants.ts
@@ -1,0 +1,8 @@
+export const TOOLTIP_TEXTS = {
+  FLOOR: 'Lowest listing price on marketplaces, excluding taker royalties and fees',
+  TOP_OFFER: 'Highest offer among all lenders providing liquidity for this collection',
+  IN_LOANS: 'Liquidity that is locked in active loans',
+  IN_OFFERS: 'Total liquidity currently available in pending offers',
+  MAX_APR:
+    'Maximum annual interest rate. Ranges between 34-104% APR depending on the loan-to-value (LTV) offered, and becomes fixed once offer is taken',
+}
