+++ src/pages/nftLending/OffersPage/components/OffersTabContent/components/HeaderList.tsx
@@ -4,8 +4,8 @@ import { TOOLTIP_TEXTS } from '../constants'
 
 import styles from '../OffersTabContent.module.less'
 
-export const HeaderTokensList = () => {
-  const tokenStats = [
+export const HeaderList = () => {
+  const stats = [
     { label: 'Floor', tooltipText: TOOLTIP_TEXTS.FLOOR },
     { label: 'Top offer', tooltipText: TOOLTIP_TEXTS.TOP_OFFER },
     { label: 'In offer', tooltipText: TOOLTIP_TEXTS.IN_OFFER },
@@ -14,15 +14,15 @@ export const HeaderTokensList = () => {
   ]
 
   return (
-    <div className={styles.headerTokensList}>
-      <div className={styles.headerTokenMainStat}>
-        <span className={styles.headerTokenStatLabel}>Collection</span>
+    <div className={styles.headerList}>
+      <div className={styles.headerMainStat}>
+        <span className={styles.headerStatLabel}>Collection</span>
       </div>
 
-      <div className={styles.headerTokenStats}>
-        {tokenStats.map(({ label, tooltipText }, index) => (
-          <div key={index} className={styles.headerTokenAdditionalStat}>
-            <span className={styles.headerTokenStatLabel}>{label}</span>
+      <div className={styles.headerStats}>
+        {stats.map(({ label, tooltipText }, index) => (
+          <div key={index} className={styles.headerAdditionalStat}>
+            <span className={styles.headerStatLabel}>{label}</span>
             {tooltipText && <Tooltip title={tooltipText} />}
           </div>
         ))}
